requirements.txt
langchain
openai
duckduckgo-search
python-dotenv
